Add a very simple fuzz test for HttpTemplate::Parse
This commit includes both a simple fuzz test as well as the
infrastructure needed to make the fuzz test run.

To run the fuzz test, use a command like:

$bazel run --config=asan-libfuzzer //test:http_template_fuzz_test_run \
    -- --clean --timeout_secs=30

As for setting up the fuzz testing framework, I followed the
directions from
https://github.com/bazelbuild/rules_fuzzing/blob/master/README.md.  At
the time of this commit, the latest version of the fuzz testing
library was release
0.1.1 (https://github.com/bazelbuild/rules_fuzzing/releases/tag/v0.1.1),
so that’s where I pointed the bazel rules.